consolidates stable mod in root
